Moved the runner action into the installed_actions setup
